@@ -2,21 +2,20 @@
 	import type { PageServerData } from './$types';
 	import { goto } from '$app/navigation';
 	import Image from '$lib/components/Image.svelte';
-	import { Button } from '$lib/components/ui/button';
-	import { writable } from 'svelte/store';
+	import { Button } from '$lib/components/ui/button/index';
 	import { format } from 'timeago.js';
 	import Status from '$lib/components/Status.svelte';
 	import Options from '$lib/components/Options.svelte';
 	import { formatPrice } from '$lib/utils/formatprice';
-  import { page } from '$app/stores';
+  import { page } from '$app/state';
 
-	export let data: PageServerData;
+	let { data }: { data: PageServerData } = $props()
 
-  let shoppingCompletedStore = writable(data.shoppingList || []);
-	let metaStore = writable(data.meta || {});
+  let shoppingCompletedStore = $state(data.shoppingList || []);
+	let metaStore = $state(data.meta || {});
 
 	// Obtener url del servidor
-	let serverUrl: string;
+	let serverUrl = $state<string>('');
 	async function getServerUrl() {
 		try {
 			const response = await fetch(`/api/server`);
@@ -28,12 +27,14 @@
 		}
 	}
 
-	async function loadSales(page: number, limit: number = 10) {
+	async function loadSales(pageNumber: number, limit: number = 10) {
 		try {
-			await getServerUrl();
+      if (!serverUrl) {
+        await getServerUrl();
+      }
 
 			const response = await fetch(
-				`${serverUrl}/users/shoppingcompleted/${$page.data.user._id}?page=${page}&limit=${limit}`
+				`${serverUrl}/users/shoppingcompleted/${page.data.user._id}?page=${pageNumber}&limit=${limit}`
 			);
 			const result = await response.json();
 
@@ -45,8 +46,8 @@
 			);
 
 			// devuelve { data: [], meta: {} }
-			shoppingCompletedStore.set(shoppingList);
-			metaStore.set(result.meta);
+			shoppingCompletedStore = shoppingList
+			metaStore = result.meta
 		} catch (error) {
 			console.error('Error al cargar los productos del usuario: ', error);
 		}
@@ -57,13 +58,17 @@
 		loadSales(newPage, 10);
 	}
 
-	$: console.log({ shoppingData: $shoppingCompletedStore });
+	$inspect({ shoppingData: shoppingCompletedStore });
 </script>
 
 <div class="flex p-5">
 	<button
 		class="flex justify-center items-center h-10 w-10 dark:bg-[#202020] rounded-sm hover:dark:bg-[#252525]"
-		on:click|preventDefault={() => goto('/shopping')}
+		onclick={(e) => {
+      e.preventDefault()
+      goto('/shopping')
+    }}
+    aria-label="Volver a Compras"
 	>
 		<iconify-icon
 			icon="material-symbols:chevron-left-rounded"
@@ -80,7 +85,7 @@
 	</div>
 </div>
 
-{#if Array.isArray($shoppingCompletedStore) && $shoppingCompletedStore.length > 0}
+{#if Array.isArray(shoppingCompletedStore) && shoppingCompletedStore.length > 0}
 	<div class="overflow-x-auto w-full p-4">
 		<table class="w-full border-collapse text-left text-sm">
 			<thead>
@@ -97,7 +102,7 @@
 				</tr>
 			</thead>
 			<tbody class="divide-y divide-gray-200 dark:divide-[#252525]">
-				{#each $shoppingCompletedStore as order}
+				{#each shoppingCompletedStore as order}
 					<tr class="hover:bg-gray-50 dark:hover:bg-[#2c2c2c]">
 						<!-- Imágenes (mostrar la primera o un recuento) -->
 						<td class="py-2 px-4 dark:text-gray-200">
@@ -153,7 +158,7 @@
 	<div class="flex justify-between mt-2 m-5">
 		<div class="">
 			<h3 class="text-sm dark:text-[#707070]">
-				Items: {$metaStore.itemCount} | Página: {$metaStore.page} de {$metaStore.pageCount}
+				Items: {metaStore.itemCount} | Página: {metaStore.page} de {metaStore.pageCount}
 			</h3>
 		</div>
 
@@ -162,8 +167,11 @@
 				class="border-gray-400 dark:border-[#252525] dark:hover:bg-[#252525]"
 				variant="outline"
 				size="sm"
-				disabled={!$metaStore.hasPreviousPage}
-				on:click={() => changePage(Number($metaStore.page) - 1)}
+				disabled={!metaStore.hasPreviousPage}
+				onclick={(e: any) => {
+          e.preventDefault()
+          changePage(Number(metaStore.page) - 1)
+        }}
 			>
 				Anterior
 			</Button>
@@ -171,8 +179,11 @@
 				class="border-gray-400 dark:border-[#252525] dark:hover:bg-[#252525]"
 				variant="outline"
 				size="sm"
-				disabled={!$metaStore.hasNextPage}
-				on:click={() => changePage(Number($metaStore.page) + 1)}
+				disabled={!metaStore.hasNextPage}
+				onclick={(e: any) => {
+          e.preventDefault()
+          changePage(Number(metaStore.page) + 1)
+        }}
 			>
 				Siguiente
 			</Button>
@@ -185,7 +196,7 @@
 			height="5rem"
 			width="5rem"
 			class="text-[#707070] mb-4"
-		/>
+		></iconify-icon>
 
 		<h1 class="text-xl font-semibold text-[#707070] mb-2">No hay productos en el historial</h1>
 		<p class="text-lg text-[#707070]">¡Explora y encuentra productos que te interesen!</p>
