@@ -60,7 +60,7 @@
 			height="5rem"
 			width="5rem"
 			class="text-[#707070] mb-4"
-		/>
+		></iconify-icon>
 		<p class="text-lg text-[#707070] mb-2">No hay productos que coincidan con tu búsqueda: "{$search}"</p>
 		<p class="text-lg text-[#707070]">
 			Intenta con otros términos o categorías para encontrar lo que buscas.
