@@ -1,6 +1,6 @@
 <script lang="ts">
 	import * as DropdownMenu from '$lib/components/ui/dropdown-menu'
-	import { Button } from '$lib/components/ui/button'
+	import { Button } from '$lib/components/ui/button/index'
 	import { MoreHorizontal } from 'lucide-svelte'
 	import { toast } from 'svelte-sonner'
 
@@ -61,7 +61,7 @@
 							height="1.1rem"
 							width="1.1rem"
 							class="text-gray-200 flex justify-center items-center"
-						/>
+						></iconify-icon>
 						<span class="ml-3">Enviando</span>
 					</DropdownMenu.Item>
 					<!-- <DropdownMenu.Separator />
